cambios rutas reservas admin

+++ routes/web.php
@@ -88,28 +88,17 @@ Route::middleware([
          Route::delete('admin/clientes/{cliente}', [App\Http\Controllers\Admin\UserController::class, 'eliminar'])->name('admin.clientes.eliminar');
 
          // Rutas para el CRUD de reservas (admin)
-         //Route::prefix('admin/reservas')->group(function () {
-             Route::get('/', [App\Http\Controllers\Admin\ReservaController::class, 'index'])->name('admin.reservas.listar');
-             Route::get('/crear', [App\Http\Controllers\Admin\ReservaController::class, 'create'])->name('admin.reservas.crear');
-             Route::post('/', [App\Http\Controllers\Admin\ReservaController::class, 'store'])->name('admin.reservas.almacenar');
-             Route::get('/{reserva}/editar', [App\Http\Controllers\Admin\ReservaController::class, 'edit'])->name('admin.reservas.editar');
-             Route::put('/{reserva}', [App\Http\Controllers\Admin\ReservaController::class, 'update'])->name('admin.reservas.actualizar');
-             Route::delete('/{reserva}', [App\Http\Controllers\Admin\ReservaController::class, 'destroy'])->name('admin.reservas.eliminar');
-
-             // AJAX
-             Route::get('/dias-disponibles', [App\Http\Controllers\Admin\ReservaController::class, 'diasDisponibles'])->name('admin.reservas.diasDisponibles');
-             Route::get('/huecos-disponibles', [App\Http\Controllers\Admin\ReservaController::class, 'huecosDisponibles'])->name('admin.reservas.huecosDisponibles');
-        // });
-         //Route::get('admin/reservas', [App\Http\Controllers\Admin\ReservaController::class, 'index'])->name('admin.reservas.listar');
-         //Route::get('admin/reservas/crear', [App\Http\Controllers\Admin\ReservaController::class, 'create'])->name('admin.reservas.crear');
-         //Route::post('admin/reservas', [App\Http\Controllers\Admin\ReservaController::class, 'store'])->name('admin.reservas.almacenar');
-         //Route::get('admin/reservas/{reserva}/editar', [App\Http\Controllers\Admin\ReservaController::class, 'edit'])->name('admin.reservas.editar');
-         //Route::put('admin/reservas/{reserva}', [App\Http\Controllers\Admin\ReservaController::class, 'update'])->name('admin.reservas.actualizar');
-         //Route::delete('admin/reservas/{reserva}', [App\Http\Controllers\Admin\ReservaController::class, 'destroy'])->name('admin.reservas.eliminar');
+         
+         Route::get('admin/reservas', [App\Http\Controllers\Admin\ReservaController::class, 'index'])->name('admin.reservas.listar');
+         Route::get('admin/reservas/crear', [App\Http\Controllers\Admin\ReservaController::class, 'create'])->name('admin.reservas.crear');
+         Route::post('admin/reservas', [App\Http\Controllers\Admin\ReservaController::class, 'store'])->name('admin.reservas.almacenar');
+         Route::get('admin/reservas/{reserva}/editar', [App\Http\Controllers\Admin\ReservaController::class, 'edit'])->name('admin.reservas.editar');
+         Route::put('admin/reservas/{reserva}', [App\Http\Controllers\Admin\ReservaController::class, 'update'])->name('admin.reservas.actualizar');
+         Route::delete('admin/reservas/{reserva}', [App\Http\Controllers\Admin\ReservaController::class, 'destroy'])->name('admin.reservas.eliminar');
 
          // Rutas para AJAX de reservas admin (días y huecos disponibles)
-         //Route::get('admin/reservas/dias-disponibles', [App\Http\Controllers\Admin\ReservaController::class, 'diasDisponibles'])->name('admin.reservas.diasDisponibles');
-         //Route::get('admin/reservas/huecos-disponibles', [App\Http\Controllers\Admin\ReservaController::class, 'huecosDisponibles'])->name('admin.reservas.huecosDisponibles');
+         Route::get('admin/reservas/dias-disponibles', [App\Http\Controllers\Admin\ReservaController::class, 'diasDisponibles'])->name('admin.reservas.diasDisponibles');
+         Route::get('admin/reservas/huecos-disponibles', [App\Http\Controllers\Admin\ReservaController::class, 'huecosDisponibles'])->name('admin.reservas.huecosDisponibles');
 
          // Rutas para el CRUD de productos (admin)
          Route::get('admin/productos', [App\Http\Controllers\Admin\ProductoController::class, 'listar'])->name('admin.productos.listar');
